sonogen: pass options as a struct
Rather than declaring equivalent member variables in both the (former)
AppOptions and App classes and passing the options one by one to the equivalent
App member variables, just pass the whole ball of options as a struct.

We can't just pass the OptionGroup to the App class because apparently copying
the OptionGroup class is illegal and causes a segfault on destruction, so just
create a little container struct to hold the option values.